requirements.txt
numpy>=1.24.0
torch>=2.0.0
requests>=2.28.0
websocket-client>=1.5.0
